+++ wfc_dylib/ffi/wfc.hpp
@@ -28,8 +28,8 @@ using Wfc = WfcState*;
 
 struct AdjacencyRule {
   AdjacencyRuleKind kind;
-  uint32_t module_low;
-  uint32_t module_high;
+  uint8_t module_low;
+  uint8_t module_high;
 };
 
 extern "C" {
